+++ app/src/main/kotlin/com/simplemobiletools/gallery/activities/EditActivity.kt
@@ -9,7 +9,7 @@ import android.graphics.Point
 import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
-import android.support.v7.widget.LinearLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.RelativeLayout