+++ app/src/main/kotlin/com/simplemobiletools/gallery/databases/GalleryDatabase.kt
@@ -1,8 +1,8 @@
 package com.simplemobiletools.gallery.databases
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
 import android.content.Context
 import com.simplemobiletools.gallery.interfaces.DirectoryDao
 import com.simplemobiletools.gallery.interfaces.MediumDao