+++ app/src/main/kotlin/com/simplemobiletools/gallery/interfaces/MediumDao.kt
@@ -1,10 +1,10 @@
 package com.simplemobiletools.gallery.interfaces
 
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Delete
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy.REPLACE
-import android.arch.persistence.room.Query
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy.REPLACE
+import androidx.room.Query
 import com.simplemobiletools.gallery.models.Medium
 
 @Dao